refactor: consolidate the website guest surface into frappe/website/api.py
Concentrate the deliberate guest-accessible endpoints - web form
accept/delete/list/data, website comments, discussions, page view
tracking, contact form, help article feedback, personal data deletion
confirmation and email unsubscribe - into one reviewable module. All of
them are marked @frappe.public(group="Website") with full annotations
and Sphinx docstrings.

Over-generic names got proper ones in the new module (accept ->
accept_web_form, delete -> delete_web_form_document, send_message ->
send_contact_message, add_feedback -> add_help_article_feedback,
delete_message -> delete_discussion_reply, confirm_deletion ->
confirm_personal_data_deletion, get_form_data -> get_web_form_data).

Old dotted paths - including the ones baked into already-sent signed
email links (unsubscribe, deletion confirmation) - keep working through
lazy module __getattr__ aliases in the original modules; the aliases
resolve to the same function objects, so whitelist identity, guest
access and rate limits are preserved.

Not moved (deliberately): OAuth2 endpoints, realtime handlers, login and
signup/password endpoints - those belong to the auth consolidation.
