+++ frappe/website/doctype/discussion_topic/discussion_topic.py
@@ -24,41 +24,22 @@ class DiscussionTopic(Document):
 	pass
 
 
-@frappe.whitelist()
-def submit_discussion(
-	doctype: str,
-	docname: str | int,
-	reply: str,
-	title: str,
-	topic_name: str | None = None,
-	reply_name: str | None = None,
-):
-	if reply_name:
-		doc = frappe.get_doc("Discussion Reply", reply_name)
-		if doc.owner != frappe.session.user:
-			frappe.throw(frappe._("You can only edit your own replies."), frappe.PermissionError)
-		doc.reply = reply
-		doc.save(ignore_permissions=True)
-		return
-
-	if topic_name:
-		save_message(reply, topic_name)
-		return topic_name
-
-	topic = frappe.get_doc(
-		{
-			"doctype": "Discussion Topic",
-			"title": title,
-			"reference_doctype": doctype,
-			"reference_docname": docname,
-		}
-	)
-	topic.save(ignore_permissions=True)
-	save_message(reply, topic.name)
-	return topic.name
-
-
 def save_message(reply, topic):
 	frappe.get_doc({"doctype": "Discussion Reply", "reply": reply, "topic": topic}).save(
 		ignore_permissions=True
 	)
+
+
+# `submit_discussion` moved to frappe.website.api. The aliases keep the old
+# dotted paths working; resolved lazily to avoid circular imports.
+_MOVED_TO_WEBSITE_API = {
+	"submit_discussion": "submit_discussion",
+}
+
+
+def __getattr__(name: str):
+	if new_name := _MOVED_TO_WEBSITE_API.get(name):
+		from frappe.website import api
+
+		return getattr(api, new_name)
+	raise AttributeError(f"module {__name__!r} has no attribute {name!r}")
